Create main.yaml
Let me explain what each section does:

name: This is just a name for the pipeline. You can call it whatever you like.
on: This specifies the events that should trigger the pipeline to run. In this case, we're triggering it on any push to the repository or any pull request to the repository.
jobs: This is where we define the jobs that make up the pipeline. In this case, we only have one job called build.
runs-on: This specifies the environment in which the job should run. In this case, we're using an Ubuntu environment.
steps: These are the individual steps that make up the job.
uses: This specifies an action that should be used in the step. Actions are reusable units of code that perform specific tasks.
with: This specifies any inputs that should be passed to the action.
run: This runs a command on the environment.
In this example, we're checking out the code, setting up Node.js, installing dependencies, and running tests. You can modify this script to suit your own needs.